# perf: optimize compute proof