--- a/.gitignore
+++ b/.gitignore
@@ -2,3 +2,8 @@ target
 Cargo.lock
 consensus-spec-tests
 cobertura.xml
+
+flamegraph.svg
+
+# macOS
+.DS_Store