--- a/Cargo.toml
+++ b/Cargo.toml
@@ -6,3 +6,6 @@ members = [
     "ssz-rs-derive",
     "ssz-rs-test-gen"
 ]
+
+[profile.bench]
+debug = true