--- a/ssz-rs/Cargo.toml
+++ b/ssz-rs/Cargo.toml
@@ -28,6 +28,7 @@ serde = { version = "1.0", default-features = false, features = [
 alloy-primitives = { version = "~0.8", default-features = false }
 
 [dev-dependencies]
+criterion = { version = "0.5", features = ["html_reports"] }
 snap = "1.0"
 project-root = "0.2.2"
 serde_json = "1.0.81"
@@ -35,3 +36,7 @@ hex = "0.4.3"
 
 [build-dependencies]
 sha2 = "0.9.8"
+
+[[bench]]
+name = "compute_proof"
+harness = false