--- a/ssz-rs/src/merkleization/mod.rs
+++ b/ssz-rs/src/merkleization/mod.rs
@@ -1,4 +1,5 @@
 pub mod generalized_index;
+pub mod hasher;
 mod merkleize;
 pub mod multiproofs;
 mod node;