Simpler alternative to the other PRs

- Use a single in-place buffer to allocate nodes
- Hash pairs in bulk
- Only hash "real" nodes

Before
```
➜  ssz-rs git:(5170755) ✗ RUSTFLAGS="-C force-frame-pointers=yes" cargo bench --bench compute_proof
Prove Benchmark - File: benches/21315748.json - size 247/123
                        time:   [455.34 ms 459.17 ms 464.71 ms]
                        change: [-4.1907% -3.1390% -1.6963%] (p = 0.00 < 0.05)
                        Performance has improved.
Prove Benchmark - File: benches/21327802.json - size 261/130
                        time:   [459.60 ms 461.11 ms 463.13 ms]
                        change: [-3.2592% -2.8113% -2.3232%] (p = 0.00 < 0.05)
                        Performance has improved.
```

After
```
➜  ssz-rs git:(hash-pairs-bulk) ✗ RUSTFLAGS="-C force-frame-pointers=yes" cargo bench --bench compute_proof --features hashtree,sha2-asm
Prove Benchmark - File: benches/21315748.json - size 247/123
                        time:   [6.3544 ms 6.4716 ms 6.5715 ms]
                        change: [-98.647% -98.616% -98.583%] (p = 0.00 < 0.05)
                        Performance has improved.

Prove Benchmark - File: benches/21327802.json - size 261/130
                        time:   [6.5145 ms 6.5672 ms 6.6093 ms]
                        change: [-98.584% -98.571% -98.555%] (p = 0.00 < 0.05)
                        Performance has improved.
```