@estensen 2025-08-06
After filling zeros more efficiently (41ad274e2fb1e91b96b847144939456747bc612b)
```
Prove Benchmark - File: benches/21315748.json - size 247/123
                        time:   [2.2082 ms 2.2183 ms 2.2298 ms]
                        change: [-66.258% -65.881% -65.575%] (p = 0.00 < 0.05)
                        Performance has improved.

Prove Benchmark - File: benches/21327802.json - size 261/130
                        time:   [2.4209 ms 2.4597 ms 2.5057 ms]
                        change: [-65.841% -64.761% -63.656%] (p = 0.00 < 0.05)
                        Performance has improved.
```